Refactored BnCKFwdInference::GetSolution and reverted changes in implicitgemm_ck_util.hpp
